% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectFitModel.R
\name{as.character}
\alias{as.character}
\alias{as.character,DirectFitModel-method}
\title{Convert a \code{\link{DirectFitModel}} to a String}
\arguments{
\item{x}{the object}
}
\value{
the name of the object
}
\description{
the \code{as.character} implementation for
  \code{\link{DirectFitModel}}
}
