% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectFitModel.R
\docType{class}
\name{DirectFitModel-class}
\alias{DirectFitModel-class}
\alias{DirectFitModel}
\title{A Spline Model}
\value{
an instance of \code{\link{DirectFitModel}}
}
\description{
This class holds a spline model.
}
\section{Slots}{

\describe{
\item{\code{name}}{the name of the spline approach}
}}

\seealso{
DirectFitModel.new
}
