% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectFitModel.R
\name{DirectFitModel.as.character}
\alias{DirectFitModel.as.character}
\title{Convert a \code{DirectFitModel} to a String}
\usage{
DirectFitModel.as.character(x)
}
\arguments{
\item{x}{the \code{DirectFitModel}}
}
\value{
the string
}
\description{
well, convert a \code{DirectFitModel} to a String
}
