% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{regressoR.direct.default}
\alias{regressoR.direct.default}
\title{Get the Default Direct Fitters}
\usage{
regressoR.direct.default()
}
\value{
the list of default fitters
}
\description{
Get the default fitters which base their results directly on the
  data.
}
